Split badge code for readability
